refactoring

+++ ORM/Person.cs
@@ -16,6 +16,6 @@ namespace ORM
 
         [Column("Rating", typeof (string), AllowNullable = true)]
         public int Rating { get; set; }
-        public long Id { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        public uint Id { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
     }
 }
